core/plugin/link/pyproject.toml
[project]
name = "spark-link"
version = "1.0.0"
description = "工具托管平台服务"
authors = ["iflytek"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "jsonschema>=4.22.0",
    "sqlalchemy>=2.0.30",
    "sqlmodel>=0.0.18",
    "loguru>=0.7.2",
    "pymysql>=1.1.1",
    "openapi-spec-validator>=0.7.1",
    "appdirs>=1.4.4",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
    "opentelemetry-proto>=1.22.0",
    "opentelemetry-semantic-conventions>=0.46b0,<0.47",
    "opentelemetry-exporter-opencensus>=0.46b0,<0.47",
    "opentelemetry-exporter-otlp>=1.22.0",
    "redis-py-cluster>=2.1.3",
    "xingchen-utils @ https://depend.iflytek.com:443/artifactory/pypi-private/xingchen_utils/1.0.8/xingchen_utils-1.0.8-py3-none-any.whl",
    "orjson>=3.10.15",
    "mcp==1.6.0",
    "aiohttp>=3.12.15",
    "rich>=14.1.0",
]
